Convert volume property to int of percentage
Done to save on its absurd float accuricity, potentially harming notes packets' network speed

+++ src/main/java/com/cstav/genshinstrument/util/ServerUtil.java
@@ -52,7 +52,7 @@ public class ServerUtil {
      * @param volume The volume of the sound to initiate
      */
     public static void sendPlayNotePackets(ServerPlayer player, Optional<InteractionHand> hand,
-            NoteSound sound, ResourceLocation instrumentId, int pitch, float volume) {
+            NoteSound sound, ResourceLocation instrumentId, int pitch, int volume) {
 
         sendPlayNotePackets(
             player, player.blockPosition(), hand,
@@ -75,7 +75,7 @@ public class ServerUtil {
      * @param PlayNotePacketDelegate The initiator of the {@link PlayNotePacket} to be sent
      */
     public static void sendPlayNotePackets(ServerPlayer player, BlockPos pos, Optional<InteractionHand> hand,
-            NoteSound sound, ResourceLocation instrumentId, NoteButtonIdentifier noteIdentifier, int pitch, float volume,
+            NoteSound sound, ResourceLocation instrumentId, NoteButtonIdentifier noteIdentifier, int pitch, int volume,
             PlayNotePacketDelegate notePacketDelegate) {
 
         final PlayNotePacket packet = notePacketDelegate.create(
@@ -110,7 +110,7 @@ public class ServerUtil {
      * @param pitch The pitch of the sound to initiate
      */
     public static void sendPlayNotePackets(Level level, BlockPos pos, NoteSound sound, ResourceLocation instrumentId,
-            int pitch, float volume) {
+            int pitch, int volume) {
         sendPlayNotePackets(
             level, pos, sound,
             instrumentId, new DefaultNoteButtonIdentifier(sound, pitch), pitch, volume,
@@ -130,7 +130,7 @@ public class ServerUtil {
      * @param PlayNotePacketDelegate The initiator of the {@link PlayNotePacket} to be sent
      */
     public static void sendPlayNotePackets(Level level, BlockPos pos, NoteSound sound,
-            ResourceLocation instrumentId, NoteButtonIdentifier noteIdentifier, int pitch, float volume,
+            ResourceLocation instrumentId, NoteButtonIdentifier noteIdentifier, int pitch, int volume,
             PlayNotePacketDelegate notePacketDelegate) {
 
         final PlayNotePacket packet = notePacketDelegate.create(
